+++ internal/db/db_certificates.go
@@ -9,8 +9,8 @@ import (
 )
 
 type Certificate struct {
-	CertificateID int `db:"certificate_id"`
-	IssuerID      int `db:"issuer_id"` // if the issuer id == certificate_id, then this is a self-signed certificate
+	CertificateID int64 `db:"certificate_id"`
+	IssuerID      int64 `db:"issuer_id"` // if the issuer id == certificate_id, then this is a self-signed certificate
 
 	CertificatePEM string `db:"certificate"`
 }
@@ -107,7 +107,7 @@ func (db *Database) AddCertificateChainToCertificateRequest(csrFilter CSRFilter,
 	if err != nil {
 		return errors.New("cert validation failed: " + err.Error())
 	}
-	parentID := 0
+	parentID := int64(0)
 	if isSelfSigned(certBundle) {
 		certRow := Certificate{
 			IssuerID:       0,
@@ -127,7 +127,7 @@ func (db *Database) AddCertificateChainToCertificateRequest(csrFilter CSRFilter,
 			return err
 		}
 		// Update the certificate to refer to itself
-		certRow.IssuerID = int(childID)
+		certRow.IssuerID = childID
 		stmt, err = sqlair.Prepare(fmt.Sprintf(updateCertificateStmt, db.certificatesTable), Certificate{})
 		if err != nil {
 			return err
@@ -136,7 +136,7 @@ func (db *Database) AddCertificateChainToCertificateRequest(csrFilter CSRFilter,
 		if err != nil {
 			return err
 		}
-		parentID = int(childID)
+		parentID = childID
 	} else {
 		// Otherwise, go through the certificate chain in reverse and add certs as their parents
 		for i := len(certBundle) - 1; i >= 0; i-- {
@@ -149,7 +149,7 @@ func (db *Database) AddCertificateChainToCertificateRequest(csrFilter CSRFilter,
 				return err
 			}
 			err = db.conn.Query(context.Background(), stmt, certRow).Get(&certRow)
-			childID := int64(certRow.CertificateID)
+			childID := certRow.CertificateID
 			if err == sqlair.ErrNoRows {
 				stmt, err = sqlair.Prepare(fmt.Sprintf(createCertificateStmt, db.certificatesTable), Certificate{})
 				if err != nil {
@@ -167,7 +167,7 @@ func (db *Database) AddCertificateChainToCertificateRequest(csrFilter CSRFilter,
 			} else if err != nil {
 				return err
 			}
-			parentID = int(childID)
+			parentID = childID
 		}
 	}
 	stmt, err := sqlair.Prepare(fmt.Sprintf(updateCertificateRequestStmt, db.certificateRequestsTable), CertificateRequest{})
